📦 feat(DownloadResultCommand.cs): add new command to download result file
The `DownloadResultCommand` class is added to the `BenchmarkingPortal.Bll.Features.Result.Commands` namespace. This command is used to request the download of a result file. It implements the `IRequest` interface and returns a tuple containing a `FileStream`, the file name, and the file extension. The command has a `Path` property to specify the path of the file to be downloaded.