✨ feat(CpuModelHeader.cs): add new class CpuModelHeader to represent a DTO for CpuModel entity
The CpuModelHeader class is added to the BenchmarkingPortal.Dal.Dtos namespace. It is a Data Transfer Object (DTO) that represents a simplified version of the CpuModel entity. The class has properties for Id, Value, and Name, which are populated from the corresponding properties of the CpuModel entity. This DTO will be used for transferring CpuModel data between layers of the application.