🐛: Fix argument exception message for DeleteBenchmarkCommandHandler
The change was made to improve the error message when a benchmark cannot be deleted due to it not being finished, by using the ExceptionMessage<Dal.Entities.Benchmark>.InUse instead of a hardcoded string. This makes the error message more informative and consistent with other exception messages in the application.